pyproject.toml
[project]
name = "elivroimagine"
version = "0.1.0"
description = "Voice-to-backlog tool for Elivro development - record thoughts, transcribe locally with Whisper"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Jimmy" }]

dependencies = [
    "openai-whisper>=20240930",
    "pynput>=1.7.6",
    "soundcard>=0.4.2",
    "pystray>=0.19.5",
    "Pillow>=10.0.0",
    "PyYAML>=6.0",
    "numpy>=1.24.0",
    "pygame>=2.5.0",
]

[tool.setuptools.package-data]
elivroimagine = ["sounds/*.mp3"]

[project.scripts]
elivroimagine = "elivroimagine.__main__:main"

[project.optional-dependencies]
dev = ["pytest>=7.4.0", "pytest-cov>=4.1.0", "pytest-timeout>=2.2.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["-v", "--tb=short", "--timeout=10"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
